+++ CHANGELOG.md
@@ -12,6 +12,7 @@ CHANGELOG
 - add context menu popup to contact list right-click
 - add support for exporting SQL query results as INSERT statements
 - convert linefeeds in profile fields to line break tags in contacts HTML export
+- retain sort direction on changing sort column in contacts HTML export
 - export contact list by current applied filter
 - order file types in save dialog by label
 - skip trying to download old shared content guaranteed to be unavailable